+++ packages/scaffold/src/views/w3m-approve-transaction-view/index.ts
@@ -8,6 +8,7 @@ import {
   ThemeController,
   RouterController
 } from '@web3modal/core'
+import { getW3mThemeVariables } from '@web3modal/common'
 
 // -- Variables ------------------------------------------- //
 const PAGE_HEIGHT = 400
@@ -95,9 +96,14 @@ export class W3mApproveTransactionView extends LitElement {
 
   private async syncTheme() {
     const emailConnector = ConnectorController.getEmailConnector()
+
     if (emailConnector) {
+      const themeMode = ThemeController.getSnapshot().themeMode
+      const themeVariables = ThemeController.getSnapshot().themeVariables
+
       await emailConnector.provider.syncTheme({
-        themeVariables: ThemeController.getSnapshot().themeVariables
+        themeVariables,
+        w3mThemeVariables: getW3mThemeVariables(themeVariables, themeMode)
       })
     }
   }