+++ packages/scaffold-utils/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @web3modal/scaffold-utils
 
+## 4.1.12-910a844.0
+
+### Patch Changes
+
+- refactor: sync theme with secure site
+
+- Updated dependencies []:
+  - @web3modal/core@4.1.12-910a844.0
+  - @web3modal/polyfills@4.1.12-910a844.0
+
 ## 4.1.11
 
 ### Patch Changes